+++ PythonBot/log.py
@@ -9,13 +9,15 @@ async def error(event, filename="errors"):
 async def message(message : discord.Message, action : str, number=0):
     file = open("logs/" + message.server.name + ".txt","a+")
     if action == "pic":
-        file.write(str(message.timestamp) + " | " + message.channel.name.encode("ascii", "replace").decode("ascii") + " | " + message.author.name.encode("ascii", "replace").decode("ascii") + " posted a pic, saved as " + str(number) + "\n")
-        print(str(message.timestamp) + " | " + message.server.name + " | " + message.channel.name + " | " + message.author.name + " posted a pic, saved as " + str(number))
+        text = "{} | {} | {} posted a pic, saved as {}".format(message.timestamp, message.channel.name.encode("ascii", "replace").decode("ascii"), message.author.name.encode("ascii", "replace").decode("ascii"), number)
+        file.write(text + "\n")
+        print(text)
     else:
         members = list(map(message.server.get_member, message.raw_mentions))
         cont = message.content
         for user in members:
             cont = cont.replace(user.mention, "@" + user.name)
-        file.write(str(message.timestamp) + " | " + message.channel.name.encode("ascii", "replace").decode("ascii") + " | " + message.author.name.encode("ascii", "replace").decode("ascii") + " " + action + ": " + cont.encode("ascii", "replace").decode("ascii") + "\n")
-        print(str(message.timestamp) + " | " + message.server.name + " | " + message.channel.name + " | " + message.author.name + " " + action + ": " + cont)
+        text = "{} | {} | {} | {} : {}".format(message.timestamp, message.channel.name.encode("ascii", "replace").decode("ascii"), message.author.name.encode("ascii", "replace").decode("ascii"), action, cont.encode("ascii", "replace").decode("ascii"))
+        file.write(text + "\n")
+        print(text)
     file.close
